requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-jose==3.3.0
python-dotenv==1.0.0
mysqlclient==2.2.1
